created a reuseable dropdown component and moved social sharing links into vue component
