+++ app/assets/stylesheets/components/_social.scss
@@ -14,9 +14,21 @@
 
 @mixin social-icon ($image) {
   @include social-basic($image);
-  @include responsive(margin-right, rem-calc(10), rem-calc(14), rem-calc(14));
-  @include responsive(width, rem-calc(31), rem-calc(41), rem-calc(41));
-  @include responsive(height, rem-calc(31), rem-calc(41), rem-calc(41));
+  display: none;
+
+  @include breakpoint($small) {
+    display: inline-block;
+    margin-right: rem-calc(10);
+    width: rem-calc(31);
+    height: rem-calc(31);
+  }
+
+  @include breakpoint($medium) {
+    display: inline-block;
+    margin-right: rem-calc(14);
+    width: rem-calc(41);
+    height: rem-calc(41);
+  }
 }
 
 @mixin social-icon-white ($image) {
