+++ app/javascript/components/social/SocialShare.vue
@@ -1,31 +1,43 @@
 <template>
-  <div class="social--share" :class="classes">
-    <span class="social--share social__trigger icon--share" @click="toggle"></span>
-    
-    <div class="social__target" :class="{ 'active': isActive }">
-      <slot></slot>
-    </div>
-  </div>
+ <div>
+  <a
+   v-for="(medium, index) in media"
+   :key="index"
+   :href="medium.url"
+   :class="medium.customClass"
+   target="_blank"
+  >
+  </a>
+  <popup :options="media" :classes="'social__popup'" v-if="isMobile"></popup>
+ </div>
 </template>
 
 <script>
-  export default {
-    name: 'social-share',
+import Popup from "../dropdown/Popup.vue";
 
-    props: {
-      classes: String
-    },
+export default {
+ name: "social-share",
+ components: { Popup },
+ props: {
+  media: {
+   type: Array,
+   required: true,
+  },
+ },
 
-    data () {
-      return {
-        isActive: false
-      }
-    },
+ data() {
+  return {
+   isMobile: false,
+   isActive: false
+  };
+ },
+ mounted() {
 
-    methods: {
-      toggle () {
-        this.isActive = !this.isActive
-      }
-    }
-  }
+ },
+ methods: {
+  toggle() {
+   this.isActive = !this.isActive;
+  },
+ },
+};
 </script>
