+++ app/javascript/packs/application.js
@@ -35,7 +35,7 @@ import MapInfographic from '../components/map/MapInfographic'
 import MapStatistics from '../components/map/MapStatistics'
 // import NavBurger from '../components/nav/NavBurger'
 import NavLink from '../components/nav/NavLink'
-// import SocialShare from '../components/social/SocialShare'
+import SocialShare from '../components/social/SocialShare'
 
 // create event hub and export so that it can be imported into .vue files
 export const eventHub = new Vue()
@@ -61,7 +61,7 @@ document.addEventListener('turbolinks:load', () => {
       MapStatistics,
       // NavBurger,
       NavLink,
-      // SocialShare
+      SocialShare
     }
   })
 })
