+++ app/views/partials/_topbar.html.erb
@@ -7,7 +7,6 @@
     <download 
       text="<%= t('download.text') %>" 
       event-element="Download report link"
-      download-links="<%= download_links_array %>"
     ></download>
 
     <%= render 'partials/social-sharing' %>
